requirements.txt
asgiref==3.6.0
Django==4.2.1
sqlparse==0.4.4


django-allauth~=0.48.0
Pillow~=9.4.0
django-ckeditor~=6.5.1
celery~=5.2.7